extract parsing into internal functions
