+++ ktor-network/posix/src/io/ktor/network/util/SocketAddressUtils.kt
@@ -19,3 +19,36 @@ internal fun SocketAddress.resolve(): List<NativeSocketAddress> = when (this) {
     is InetSocketAddress -> getAddressInfo(hostname, port)
     is UnixSocketAddress -> listOf(NativeUnixSocketAddress(AF_UNIX.convert(), path))
 }
+
+internal fun parseIPv4String(ipString: String): ByteArray? {
+    return try {
+        val octets = ipString.split('.', limit = 4).map { o ->
+            o.toUByte().toByte()
+        }
+
+        List(4) { i ->
+            octets.getOrElse(i) { 0 }
+        }.toByteArray()
+    } catch (_: Throwable) {
+        null
+    }
+}
+
+internal fun parseIPv6String(ipString: String): ByteArray? {
+    return try {
+        val groups = ipString.split(':', limit = 8)
+        val emptyGroups = 8 - groups.count { g -> g.isNotEmpty() }
+
+        val bytes = groups.flatMap { g ->
+            if (g.isEmpty()) {
+                List(emptyGroups * 2) { 0 }
+            } else {
+                val int = g.toInt(16)
+                listOf((int shr 8).toByte(), int.toByte())
+            }
+        }
+        List(16) { i -> bytes.getOrElse(i) { 0 } }.toByteArray()
+    } catch (_: Throwable) {
+        null
+    }
+}
